autotools: added --enable-devel, v1
The idea here is instead of having to comment out bits and pieces of
Makefile.am when we go to run through a new Ubuntu release for the
first time, configure --enable-devel to do that for you and create
es-devel.sqsh w/ compilation toolchain and w/out 3rd-party stuff (that
doesn't exist yet).

This isn't done yet, I've just plumbed up the respective commented out
blocks in Makefile.am with FIXMEs.  Need to add the configure flag and
an AM_CONDITIONAL to key off of.