split source ruckus packages into es.sqsh, vmm.sqsh, and workstation.sqsh
Now that ruckusrd's utils have all been releases as standalone
packages, we don't need to lump them all in the base es.sqsh image.
Pingenator, doit, and extractenator are now in es.sqsh because I can
see a use case for them being there.  Ruckusrd, firmwarenator, and
microcodenator go in vmm.sqsh because that's where we start to need
the ability to tweak our bootup and have actual hardware.
Kernel-builder goes in workstation, because that's where we add all
the development tools.